Converted the whole class from ints to longs

+++ src/mgi/frontend/datamodel/sort/SmartAlphaComparator.java
@@ -1,27 +1,27 @@
 package mgi.frontend.datamodel.sort;
 
-	import java.util.ArrayList;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.HashSet;
 import java.util.Set;
 
-	/*
-	 * "smart alpha" means Zfp10 comes after zfp9, instead of zfp1.
-	 * It sorts each group of alpha and numeric phrases.
-	 * 
-	 * 
-	 * This comparator works on lists of Strings, but can be subclasses to work on any complex object
-	 * 	with a string field.
-	 * 
-	 * You would just need to call super.compare(string1,string2) in the subclass compare method
-	 * 
-	 * author kstone
-	 * 
-	 */
+/*
+ * "smart alpha" means Zfp10 comes after zfp9, instead of zfp1.
+ * It sorts each group of alpha and numeric phrases.
+ * 
+ * 
+ * This comparator works on lists of Strings, but can be subclasses to work on any complex object
+ * 	with a string field.
+ * 
+ * You would just need to call super.compare(string1,string2) in the subclass compare method
+ * 
+ * author kstone
+ * 
+ */
 public class SmartAlphaComparator<T> implements Comparator<T> {
 	/* The following algorithm was taken from the symbolsort module in lib python */
-	
+
 	private Set<Character> digits = new HashSet<Character>(Arrays.asList('0','1','2','3','4','5','6','7','8','9'));
 	private ArrayList<Object> splitter (String s) {
 		ArrayList<Object> splits = new ArrayList<Object>();
@@ -53,14 +53,14 @@ public class SmartAlphaComparator<T> implements Comparator<T> {
 		//return items;
 		return splits;
 	}
-	
+
 	/*
 	 * This method should be overriden to be useful
 	 */
 	public int compare(T o1, T o2) {
 		return compare(o1.toString(),o2.toString());
 	}
-	
+
 	/* these values must really be Strings, but we want to keep the 
 	 * class level type as Object to allow for flexible subclassing
 	 */
@@ -72,7 +72,7 @@ public class SmartAlphaComparator<T> implements Comparator<T> {
 			// split them into parts first
 			ArrayList<Object> s1 = splitter(string1);
 			ArrayList<Object> s2 = splitter(string2);
-			
+
 			// go through each part until we find a difference
 			for (int i = 0; i< Math.min(s1.size(), s2.size()); i++) {
 				Object ob1 = s1.get(i);
@@ -80,13 +80,13 @@ public class SmartAlphaComparator<T> implements Comparator<T> {
 				//check instance
 				if(ob1.getClass() != ob2.getClass()) {
 					// numbers go first
-					if(ob1 instanceof Integer) return -1;
+					if(ob1 instanceof Long) return -1;
 					else return 1;
 				} else {
 					// continue on if they are equal
 					if (!ob1.equals(ob2)) {
 						// just cast the correct type and return the compare functions for Integer or String
-						if(ob1 instanceof Integer) return ((Integer)ob1).compareTo((Integer)ob2);							
+						if(ob1 instanceof Long) return ((Long)ob1).compareTo((Long)ob2);
 						else return ((String)ob1).compareToIgnoreCase((String)ob2);
 					}
 				}
